Correçao sales

+++ migrations/20241008013521_create_products_table.sql
@@ -3,6 +3,6 @@ CREATE TABLE products (
     id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
     name VARCHAR(100) NOT NULL,
     description TEXT,
-    price DECIMAL(10, 2) NOT NULL,
+    price DOUBLE PRECISION NOT NULL,
     stock_quantity INT NOT NULL
 );
